requirements.txt
flask
flask-cors
sqlalchemy
pillow
tos
volcengine
python-dotenv
pydantic
pymysql